requirements.txt
Django>=6.0
djangorestframework>=3.14
drf-spectacular>=0.27
